Update authentication
Added optional tdauser and tdapass input parameters for authentication function, to allow for fully automated oauth2 authentication if they are passed, so we don't have manually type the information into the browser.  If the optional parameters are not specified, function will still check to see if TDAUSER and TDAPASS environment variables are specified and use those for fully automated authentication instead.