Added saved_orders and orders functions.
This is a basic stub function, I don't have tests cause I didn't TDD for these.
I have additional supporting libraries that add the rest of the functionality to complete an order,
but they depend on the QuickType library that I've developed. It's too much for one commit to add all that to this
library. I need to figure out a way to integrate them into this library, and I will attempt
to add them as soon as I have time.