Refactored tdclient with session handler and exceptions
I use a design pattern described by Ben Homnick (http://bhomnick.net/design-pattern-python-api-client/) to separate the transport request from the application logic. The TDClient now uses a TDASession class, which inherits from requests. It doesn't do much right now, but my plan is to move the authentication functionality to it at some point.

The client _request function passes all the various methods to the session, and will throw exceptions if the response is anything other than a 200, or 201 as happens when placing orders. This check will probably get moved into session so that it can try refreshing the token when 401s are thrown.

I added tests for as much as I could; I'm starting to get a better handle on mock, so I think this is moving in the right direction.